feat: separated details card item
